plants/requirements.txt
flask>=3.0
apscheduler>=3.10
python-dotenv>=1.0
resend>=2.0
